requirements.txt
redis
requests
Appium-Python-Client
selenium